Upper level function for t free threshold and renaming.

+++ project2/src/support_bound.py
@@ -1,3 +1,3 @@
-def support_bound(sample_size: int) -> int:
+def support_bound(sample_size: int) -> float:
 	bound = sample_size ** (1 / 3)
 	return bound
